Refactor code from scripts to utils
