+++ utils/github_utils.py
@@ -18,6 +18,9 @@ Example:
     repo = get_repository(client, "username/repository")
     issues = get_issues(repo)
 """
+from typing import List
+import requests
+
 from github import Github
 
 # Constants for tags, priority levels, and issue types
@@ -59,3 +62,72 @@ def append_response_to_issue(issue, response: str):
     """Append the complete response to the issue comments."""
     comment = f"## OpenAI API Response\n\n{response}"
     return add_comment(issue, comment)
+
+class GitHubLabelManager:
+    """
+    Manages GitHub issue labels, ensuring required labels exist and are applied.
+    """
+    def __init__(self, github_token: str):
+        """
+        Initialize the label manager.
+        
+        Args:
+            github_token (str): GitHub authentication token.
+        """
+        self.github_token = github_token
+        self.headers = {
+            "Authorization": f"token {github_token}",
+            "Accept": "application/vnd.github.v3+json"
+        }
+    
+    def ensure_labels_exist(
+        self, 
+        repo_owner: str, 
+        repo_name: str, 
+        labels: List[str], 
+        color: str = "6f42c1"
+    ) -> None:
+        """
+        Ensure specified labels exist in the repository.
+        
+        Args:
+            repo_owner (str): GitHub repository owner.
+            repo_name (str): GitHub repository name.
+            labels (List[str]): Labels to ensure exist.
+            color (str, optional): Default color for new labels.
+        """
+        url = f"https://api.github.com/repos/{repo_owner}/{repo_name}/labels"
+        
+        # Get existing labels
+        response = requests.get(url, headers=self.headers)
+        existing_labels = [label['name'] for label in response.json()]
+        
+        # Create missing labels
+        for label in labels:
+            if label not in existing_labels:
+                label_data = {"name": label, "color": color}
+                requests.post(url, headers=self.headers, json=label_data)
+    
+    def add_labels_to_issue(
+        self, 
+        repo_owner: str, 
+        repo_name: str, 
+        issue_number: int, 
+        labels: List[str]
+    ) -> bool:
+        """
+        Add labels to a specific GitHub issue.
+        
+        Args:
+            repo_owner (str): GitHub repository owner.
+            repo_name (str): GitHub repository name.
+            issue_number (int): Issue number to label.
+            labels (List[str]): Labels to add.
+        
+        Returns:
+            bool: True if labels were successfully added, False otherwise.
+        """
+        url = f"https://api.github.com/repos/{repo_owner}/{repo_name}/issues/{issue_number}/labels"
+        response = requests.post(url, headers=self.headers, json={"labels": labels})
+        return response.status_code == 200
+
